+++ src/index.ts
@@ -19,6 +19,17 @@ const isClient = (
 
 const useIsomorphicLayoutEffect = isClient ? useLayoutEffect : useEffect;
 
+const useAbortSignal = () => {
+  const abortController = useRef(new AbortController());
+  useEffect(() => {
+    const abort = () => {
+      abortController.current.abort();
+    };
+    return abort;
+  }, []);
+  return abortController.current.signal;
+};
+
 export type AsyncActionHandlers<
   R extends Reducer<any, any>,
   AsyncAction extends { type: string }
@@ -28,6 +39,7 @@ export type AsyncActionHandlers<
   } ? (
     dispatch: Dispatch<ReducerAction<R>>,
     getState: () => ReducerState<R>,
+    signal: AbortSignal,
   ) => (a: A) => Promise<void> : never : never;
 };
 
@@ -49,12 +61,12 @@ export function useReducerAsync<
  * import { useReducerAsync } from 'use-reducer-async';
  *
  * const asyncActionHandlers = {
- *   SLEEP: (dispatch, getState) => async (action) => {
+ *   SLEEP: (dispatch, getState, abortSignal) => async (action) => {
  *     dispatch({ type: 'START_SLEEP' });
  *     await new Promise(r => setTimeout(r, action.ms));
  *     dispatch({ type: 'END_SLEEP' });
  *   },
- *   FETCH: (dispatch, getState) => async (action) => {
+ *   FETCH: (dispatch, getState, abortSignal) => async (action) => {
  *     dispatch({ type: 'START_FETCH' });
  *     try {
  *       const response = await fetch(action.url);
@@ -88,6 +100,7 @@ export function useReducerAsync<
   initializer: unknown,
   asyncActionHandlers?: AsyncActionHandlers<R, AsyncAction>,
 ): [ReducerState<R>, Dispatch<ExportAction>] {
+  const abortSignal = useAbortSignal();
   const aaHandlers = (
     asyncActionHandlers || initializer
   ) as AsyncActionHandlers<R, AsyncAction>;
@@ -108,12 +121,13 @@ export function useReducerAsync<
     ) as (typeof action extends AsyncAction ? (
       dispatch: Dispatch<ReducerAction<R>>,
       getState: () => ReducerState<R>,
+      signal: AbortSignal,
     ) => (a: typeof action) => Promise<void> : null);
     if (aaHandler) {
-      aaHandler(rawDispatch, getState)(action as AsyncAction);
+      aaHandler(rawDispatch, getState, abortSignal)(action as AsyncAction);
     } else {
       rawDispatch(action as ReducerAction<R>);
     }
-  }, [aaHandlers, getState]);
+  }, [aaHandlers, getState, abortSignal]);
   return [state, dispatch];
 }