+++ __tests__/01_basic_spec.tsx
@@ -34,7 +34,7 @@ describe('basic spec', () => {
     };
     type AsyncAction = { type: 'SLEEP'; ms: number };
     const asyncActionHandlers: AsyncActionHandlers<Reducer<State, Action>, AsyncAction> = {
-      SLEEP: (dispatch) => async (action) => {
+      SLEEP: ({ dispatch }) => async (action) => {
         dispatch({ type: 'START_SLEEP' });
         await new Promise((r) => setTimeout(r, action.ms));
         dispatch({ type: 'END_SLEEP' });