+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Improve isClient detection for useIsomorphicLayoutEffect
+- Support abortability
 
 ## [1.0.0] - 2020-02-26
 ### Changed